Align with w3c js API
