+++ example/lib/src/get_user_media_sample_web.dart
@@ -27,7 +27,7 @@ class _GetUserMediaSampleState extends State<GetUserMediaSample> {
   void initState() {
     super.initState();
     initRenderers();
-    navigator().getSources().then((md) {
+    navigator.getSources().then((md) {
       setState(() {
         cameras = md.where((d) => d['kind'] == 'videoinput').toList();
       });
@@ -62,7 +62,7 @@ class _GetUserMediaSampleState extends State<GetUserMediaSample> {
     };
 
     try {
-      var stream = await navigator().getUserMedia(mediaConstraints);
+      var stream = await navigator.getUserMedia(mediaConstraints);
       _localStream = stream;
       _localRenderer.srcObject = _localStream;
     } catch (e) {
