+++ lib/src/media_devices.dart
@@ -1,23 +1,25 @@
+import 'dart:html';
+
 import 'package:flutter_webrtc/flutter_webrtc.dart';
 
 class MediaDevices {
   @Deprecated(
-      'Use the navigator(). getUserMedia(Map<String, dynamic>) provide from the facrory instead')
+      'Use the navigator.mediaDevices.getUserMedia(Map<String, dynamic>) provide from the facrory instead')
   static Future<MediaStream> getUserMedia(
       Map<String, dynamic> mediaConstraints) async {
-    return navigator().getUserMedia(mediaConstraints);
+    return navigator.mediaDevices.getUserMedia(mediaConstraints);
   }
 
   @Deprecated(
-      'Use the navigator().getDisplayMedia(Map<String, dynamic>) provide from the facrory instead')
+      'Use the navigator.mediaDevices.getDisplayMedia(Map<String, dynamic>) provide from the facrory instead')
   static Future<MediaStream> getDisplayMedia(
       Map<String, dynamic> mediaConstraints) async {
-    return navigator().getDisplayMedia(mediaConstraints);
+    return navigator.mediaDevices.getDisplayMedia(mediaConstraints);
   }
 
   @Deprecated(
-      'Use the navigator().getSources() provide from the facrory instead')
+      'Use the navigator.mediaDevices.getSources() provide from the facrory instead')
   static Future<List<dynamic>> getSources() {
-    return navigator().getSources();
+    return navigator.mediaDevices.getSources();
   }
 }
