+++ lib/src/native/rtc_peerconnection_factory.dart
@@ -1,14 +1,15 @@
 import 'dart:async';
 
+import 'package:flutter_webrtc/src/interface/navigator.dart';
+import 'package:flutter_webrtc/src/native/navigator_impl.dart';
+
 import '../interface/factory.dart';
 import '../interface/media_recorder.dart';
 import '../interface/media_stream.dart';
-import '../interface/navigator.dart';
 import '../interface/rtc_peerconnection.dart';
 import '../interface/rtc_video_renderer.dart';
 import 'media_recorder_impl.dart';
 import 'media_stream_impl.dart';
-import 'navigator_impl.dart';
 import 'rtc_peerconnection_impl.dart';
 import 'rtc_video_view_impl.dart';
 import 'utils.dart';
@@ -24,10 +25,6 @@ Future<MediaStream> createLocalMediaStream(String label) async {
   return _RTCFactoryNative.instance.createLocalMediaStream(label);
 }
 
-Navigator navigator() {
-  return _RTCFactoryNative.instance.navigator();
-}
-
 MediaRecorder mediaRecorder() {
   return _RTCFactoryNative.instance.mediaRecorder();
 }
@@ -35,6 +32,8 @@ MediaRecorder mediaRecorder() {
 RTCVideoRenderer videoRenderer() {
   return _RTCFactoryNative.instance.videoRenderer();
 }
+
+Navigator get navigator => _RTCFactoryNative.instance.navigator;
 
 class _RTCFactoryNative extends RTCFactory {
   _RTCFactoryNative._internal();
@@ -77,11 +76,6 @@ class _RTCFactoryNative extends RTCFactory {
   }
 
   @override
-  Navigator navigator() {
-    return NavigatorNative();
-  }
-
-  @override
   MediaRecorder mediaRecorder() {
     return MediaRecorderNative();
   }
@@ -90,4 +84,7 @@ class _RTCFactoryNative extends RTCFactory {
   RTCVideoRenderer videoRenderer() {
     return RTCVideoRendererNative();
   }
+
+  @override
+  Navigator get navigator => NavigatorNative();
 }
