+++ lib/src/web/rtc_peerconnection_factory.dart
@@ -2,15 +2,16 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:html' as html;
 
+import 'package:flutter_webrtc/src/interface/navigator.dart';
+import 'package:flutter_webrtc/src/web/navigator_impl.dart';
+
 import '../interface/factory.dart';
 import '../interface/media_recorder.dart';
 import '../interface/media_stream.dart';
-import '../interface/navigator.dart';
 import '../interface/rtc_peerconnection.dart';
 import '../interface/rtc_video_renderer.dart';
 import 'media_recorder_impl.dart';
 import 'media_stream_impl.dart';
-import 'navigator_impl.dart';
 import 'rtc_peerconnection_impl.dart';
 import 'rtc_video_view_impl.dart';
 
@@ -25,9 +26,7 @@ Future<MediaStream> createLocalMediaStream(String label) {
   return _RTCFactoryWeb.instance.createLocalMediaStream(label);
 }
 
-Navigator navigator() {
-  return _RTCFactoryWeb.instance.navigator();
-}
+Navigator get navigator => _RTCFactoryWeb.instance.navigator;
 
 MediaRecorder mediaRecorder() {
   return _RTCFactoryWeb.instance.mediaRecorder();
@@ -62,11 +61,6 @@ class _RTCFactoryWeb extends RTCFactory {
   }
 
   @override
-  Navigator navigator() {
-    return NavigatorWeb();
-  }
-
-  @override
   MediaRecorder mediaRecorder() {
     return MediaRecorderWeb();
   }
@@ -75,4 +69,7 @@ class _RTCFactoryWeb extends RTCFactory {
   RTCVideoRenderer videoRenderer() {
     return RTCVideoRendererWeb();
   }
+
+  @override
+  Navigator get navigator => NavigatorWeb();
 }
