+++ test/unit/web/rtc_videw_view_test.dart
@@ -1,6 +1,6 @@
 @TestOn('browser')
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_webrtc/src/web/navigator_impl.dart';
+import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:flutter_webrtc/src/web/rtc_video_view_impl.dart';
 
 void main() {
@@ -8,7 +8,7 @@ void main() {
   test('should complete succesfully', () async {
     var renderer = RTCVideoRendererWeb();
     await renderer.initialize();
-    renderer.srcObject = await NavigatorWeb().getUserMedia({});
+    renderer.srcObject = await navigator.getUserMedia({});
     await renderer.dispose();
   });
 }
